remove cli and clippy error
